requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
langchain_openai
python-dotenv
langchain-community
pydantic
psycopg2-binary
mlflow
boto3
datasets